Separated out models into individual files per section
This makes everying much more maintainable. I'll probably do this for views eventually but I don't really feel the need to yet
