+++ un_project/un_records_app/models/__init__.py
@@ -1,0 +1,7 @@
+from .alliance_models import *
+from .charter_models import *
+from .court_case_models import *
+from .declaration_of_war_models import *
+from .executive_order_models import *
+from .resolution_models import *
+from .treaty_models import *
