+++ un_project/un_records_app/models/court_case_models.py
@@ -1,0 +1,4 @@
+from django.db import models
+from django.core.exceptions import ValidationError
+from un_app.models import Nation
+
